Add the ADD_EPH_HS command to the controller.
Adds a new ephemeral hidden service.  The syntax is:

  "ADD_EPH_HS" keytype SP key SP 1*(port SP addrPort) CRLF
   keytype = "RSA1024"
   key = the base64 encoded DER representation of a PKCS#1 format
         1024 bit RSA private key
   port = the "VIRTPORT"
   addrPort = the "TARGET"

A hidden service is created using the key and list of port/targets,
that will persist till configuration reload or the termination of the
tor process.  "DEL_EPH_HS" will be added at a later date.

Proper key management, and deriving the onion address from the public
key is left as an exercise for the application invoking the controller
command.