Limit `DEL_EPH_HS` to removing HSes originating from the current conn.
Only connections that created the eph. HS with `ADD_EPH_HS` can remove
them now, since the lifespan of the HS is tied to the originating
control connection anyway.

Technically, there's probably other ways to mess with eph. HSes via the
control port, but fixing the control protocol is a separate task.