(squash) Replace GET_ONIONS with "GETINFO onions/".
"GETINFO onions/current" returns a list of Onion Services that are
owned by the current control port connection.

"GETINFO onions/detached" returns a list of Onion Services that are
owned by no particular control port connection.

The original commit message for this feature should be altered on the
final rebase/squash to mention this.